requirements.txt
supabase==2.28.3
yfinance==0.2.36
pandas==2.2.0
python-dotenv==1.0.1
pandas-ta==0.3.14b0

# For Stage 4 Telegram & Heatmap + Stage 5 Dashboard
matplotlib
seaborn
requests
streamlit
